Use `i64::from` instead of `u32 as i64`
...because it guarantees that the conversion is lossless

+++ core/lib/dal/src/events_web3_dal.rs
@@ -127,9 +127,8 @@ impl EventsWeb3Dal<'_, '_> {
     fn build_get_logs_where_clause(&self, filter: &GetLogsFilter) -> (String, u8) {
         let mut arg_index = 1;
 
-        let mut where_sql = format!("(miniblock_number >= {})", filter.from_block.0 as i64);
-
-        where_sql += &format!(" AND (miniblock_number <= {})", filter.to_block.0 as i64);
+        let mut where_sql = format!("(miniblock_number >= {})", filter.from_block.0);
+        where_sql += &format!(" AND (miniblock_number <= {})", filter.to_block.0);
 
         // Add filters for address (like `address = ANY($1)` or `address = $1`)
         if let Some(filter_sql) =
